Move DSClientSession Class from Base class
Remove BaseDSClientSession
