+++ PSAsigraDSClient.csproj
@@ -79,7 +79,6 @@
     <Compile Include="PSAsigraDSClient\BaseDSClientRetentionRuleParams.cs" />
     <Compile Include="PSAsigraDSClient\BaseDSClientScheduleDetail.cs" />
     <Compile Include="PSAsigraDSClient\BaseDSClientScheduleDetailParams.cs" />
-    <Compile Include="PSAsigraDSClient\BaseDSClientSession.cs" />
     <Compile Include="PSAsigraDSClient\BaseDSClientSessionCleanup.cs" />
     <Compile Include="PSAsigraDSClient\BaseDSClientSetScheduleDetail.cs" />
     <Compile Include="PSAsigraDSClient\BaseDSClientStartBackupSetActivity.cs" />
@@ -92,6 +91,7 @@
     <Compile Include="PSAsigraDSClient\DisableDSClientBackupSet.cs" />
     <Compile Include="PSAsigraDSClient\DisableDSClientSchedule.cs" />
     <Compile Include="PSAsigraDSClient\DisconnectDSClientSession.cs" />
+    <Compile Include="PSAsigraDSClient\DSClientSession.cs" />
     <Compile Include="PSAsigraDSClient\EnableDSClientBackupSet.cs" />
     <Compile Include="PSAsigraDSClient\EnableDSClientSchedule.cs" />
     <Compile Include="PSAsigraDSClient\ExportDSClientConfig.cs" />
